+++ src/collective/autoscaling/setuphandlers.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+
+from Products.CMFPlone.interfaces import INonInstallable
+from zope.interface import implementer
+
+
+@implementer(INonInstallable)
+class HiddenProfiles(object):
+    def getNonInstallableProfiles(self):
+        """Hide unwanted profiles from site-creation and quickinstaller."""
+        return [
+            "collective.autoscaling:testing",
+            "collective.autoscaling:uninstall",
+        ]